Extracting ActivityWriter to minimize public API
