+++ src/SerilogTracing.OpenTelemetry.Exporter/SerilogTraceExporter.cs
@@ -1,27 +1,13 @@
-﻿using SerilogTracing.Instrumentation;
-using SerilogTracing.Interop;
-
-namespace SerilogTracing.OpenTelemetry.Exporter;
+﻿namespace SerilogTracing.OpenTelemetry.Exporter;
 
 internal class SerilogTraceExporter(ILogger logger) : BaseExporter<Activity>
 {
+    private readonly ActivityWriter _writer = new(() => logger);
+
     public override ExportResult Export(in Batch<Activity> batch)
     {
         foreach (var activity in batch)
-        {
-            if (ActivityInstrumentation.HasAttachedLoggerActivity(activity))
-                continue; // `LoggerActivity` completion writes these to the activity-specific logger.
-
-            var activityLogger = !string.IsNullOrWhiteSpace(activity.Source.Name)
-                ? logger.ForContext(Serilog.Core.Constants.SourceContextPropertyName, activity.Source.Name)
-                : logger;
-
-            var level = ActivityInstrumentation.GetCompletionLevel(activity);
-            if (!activityLogger.IsEnabled(level))
-                continue;
-
-            activityLogger.Write(ActivityConvert.ActivityToLogEvent(activityLogger, activity));
-        }
+            _writer.Write(activity);
 
         return ExportResult.Success;
     }
